+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Releases
 
+- **Added:** Option `--statistics`: to show counts for every rule with at least one violation.
+- **Fixed:** Bug in walking directories when using `--ignore` option.
+
 ### v0.0.3 (2025-09-18)
 
-- **Added:** Arguments "--enable" and "--disable" to control specific checks.
+- **Added:** Options `--enable` and `--disable` to control specific checks.
 - **Changed:** Update README.md to include installation instructions.
 
 ### v0.0.2 (2025-09-17)